Factorise diff de versions
